backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
firebase-admin==6.6.0
anthropic==0.44.0
google-cloud-secret-manager==2.21.1
google-cloud-storage==2.19.0
google-cloud-firestore==2.19.0
pypdf==5.1.0
python-docx==1.1.2
defusedxml==0.7.1
httpx==0.28.1
structlog==24.4.0
pydantic==2.10.4
pydantic-settings==2.7.1
websockets==14.1
